+++ android/koin-android/src/main/java/org/koin/androidx/viewmodel/ViewModelOwner.kt
@@ -3,17 +3,27 @@ package org.koin.androidx.viewmodel
 import androidx.lifecycle.ViewModelStoreOwner
 import androidx.savedstate.SavedStateRegistryOwner
 
-//TODO Deprecate in 3.2
+
+typealias ViewModelStoreOwnerProducer = () -> ViewModelStoreOwner
+
+@Deprecated("Replaced by ViewModelStoreOwnerProducer")
 typealias ViewModelOwnerDefinition = () -> ViewModelOwner
+
+@Deprecated("Replaced by ViewModelStoreOwner")
 class ViewModelOwner(
-        val storeOwner: ViewModelStoreOwner,
-        val stateRegistry: SavedStateRegistryOwner? = null,
+    val storeOwner: ViewModelStoreOwner,
+    val stateRegistry: SavedStateRegistryOwner? = null,
 ) {
     companion object {
-        fun from(storeOwner: ViewModelStoreOwner, stateRegistry: SavedStateRegistryOwner? = null) = ViewModelOwner(storeOwner,stateRegistry)
+        @Deprecated("Replaced by ViewModelStoreOwner")
+        fun from(storeOwner: ViewModelStoreOwner, stateRegistry: SavedStateRegistryOwner? = null) =
+            ViewModelOwner(storeOwner, stateRegistry)
 
+        @Deprecated("Replaced by ViewModelStoreOwner")
         fun from(storeOwner: ViewModelStoreOwner) = ViewModelOwner(storeOwner)
 
-        fun fromAny(owner: Any) = ViewModelOwner((owner as ViewModelStoreOwner), owner as? SavedStateRegistryOwner)
+        @Deprecated("Replaced by ViewModelStoreOwner")
+        fun fromAny(owner: Any) =
+            ViewModelOwner((owner as ViewModelStoreOwner), owner as? SavedStateRegistryOwner)
     }
 }