+++ android/koin-android/src/main/java/org/koin/androidx/viewmodel/ext/android/ActivityVM.kt
@@ -19,10 +19,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.viewModels
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelStoreOwner
-import androidx.savedstate.SavedStateRegistryOwner
 import org.koin.android.ext.android.getKoinScope
-import org.koin.androidx.viewmodel.ViewModelOwner
-import org.koin.androidx.viewmodel.ViewModelOwnerDefinition
 import org.koin.core.annotation.KoinInternalApi
 import org.koin.core.parameter.ParametersDefinition
 import org.koin.core.qualifier.Qualifier
@@ -38,8 +35,7 @@ import org.koin.core.qualifier.Qualifier
 @OptIn(KoinInternalApi::class)
 inline fun <reified T : ViewModel> ComponentActivity.viewModel(
         qualifier: Qualifier? = null,
-        //TODO Not needed - just internal
-        noinline owner : ViewModelOwnerDefinition = { ViewModelOwner.from(this as ViewModelStoreOwner, this as? SavedStateRegistryOwner) },
+        owner : ViewModelStoreOwner = this,
         noinline parameters: ParametersDefinition? = null
 ): Lazy<T> {
         val scope = getKoinScope()
@@ -50,8 +46,7 @@ inline fun <reified T : ViewModel> ComponentActivity.viewModel(
 
 inline fun <reified T : ViewModel> ComponentActivity.getViewModel(
         qualifier: Qualifier? = null,
-        //TODO Not needed - just internal
-        noinline owner : ViewModelOwnerDefinition = { ViewModelOwner.from(this as ViewModelStoreOwner, this as? SavedStateRegistryOwner) },
+        owner : ViewModelStoreOwner = this,
         noinline parameters: ParametersDefinition? = null,
 ): T {
         return viewModel<T>(qualifier, owner, parameters).value